# backend/cmd/server/main.go
package main

import (
	"context"
	"fmt"
	"log"
	"net/http"
	"os"
	"os/signal"
	"strings"
	"syscall"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/joho/godotenv"
	"gorm.io/driver/postgres"
	"gorm.io/gorm"
	"gorm.io/gorm/logger"

	"tano_blog/backend/internal/config"
	"tano_blog/backend/internal/handler"
	"tano_blog/backend/internal/middleware"
	"tano_blog/backend/internal/model"
	"tano_blog/backend/internal/repository"
	"tano_blog/backend/internal/service"
	"tano_blog/backend/internal/utils"
)

func main() {
	cfg := config.Load()

	// Connect to database
	logLevel := logger.Warn
	if gin.Mode() == gin.DebugMode {
		logLevel = logger.Info
	}
	db, err := gorm.Open(postgres.Open(cfg.Database.DSN), &gorm.Config{
		Logger: logger.Default.LogMode(logLevel),
	})
	if err != nil {
		log.Fatalf("Failed to connect to database: %v", err)
	}
	log.Println("Connected to database")

	// Auto migrate
	if err := model.AutoMigrate(db); err != nil {
		log.Fatalf("Failed to migrate database: %v", err)
	}
	log.Println("Database migrated")

	// Enable pg_trgm extension for full-text search
	db.Exec("CREATE EXTENSION IF NOT EXISTS pg_trgm")
	db.Exec("CREATE INDEX IF NOT EXISTS idx_posts_title_trgm ON posts USING GIN (title gin_trgm_ops)")
	db.Exec("CREATE INDEX IF NOT EXISTS idx_posts_content_trgm ON posts USING GIN (content gin_trgm_ops)")

	// Seed default admin user if not exists
	seedAdmin(db, cfg.AdminPassword)

	// Seed default configs
	seedSiteConfigs(db)

	// Initialize repositories
	postRepo := repository.NewPostRepo(db)
	commentRepo := repository.NewCommentRepo(db)
	mediaRepo := repository.NewMediaRepo(db)
	seriesRepo := repository.NewSeriesRepo(db)

	// Initialize services
	emailService := service.NewEmailService(db)

	// Initialize handlers
	authHandler := handler.NewAuthHandler(db, &cfg.JWT, emailService)
	postHandler := handler.NewPostHandler(postRepo)
	categoryHandler := handler.NewCategoryHandler(db)
	tagHandler := handler.NewTagHandler(db)
	commentHandler := handler.NewCommentHandler(commentRepo, db, emailService)
	mediaHandler := handler.NewMediaHandler(mediaRepo, &cfg.Upload)
	siteConfigHandler := handler.NewSiteConfigHandler(db, emailService)
	accessLogHandler := handler.NewAccessLogHandler(db)
	backupHandler := handler.NewBackupHandler(db, cfg.Upload.Dir, cfg.BackupDir)
	seriesHandler := handler.NewSeriesHandler(seriesRepo)
	feedHandler := handler.NewFeedHandler(db)
	friendLinkHandler := handler.NewFriendLinkHandler(db)
	navLinkHandler := handler.NewNavLinkHandler(db)

	// Setup router
	r := gin.New()
	r.SetTrustedProxies(nil)
	r.Use(gin.Recovery())
	r.Use(middleware.AccessLogger(db))

	// CORS: allow frontend origin
	allowedOrigins := strings.Split(os.Getenv("CORS_ORIGINS"), ",")
	if len(allowedOrigins) == 1 && allowedOrigins[0] == "" {
		allowedOrigins = []string{"http://localhost:3000"}
	}
	r.Use(middleware.CORS(allowedOrigins))
	r.Use(middleware.SecurityHeaders())

	// Serve uploaded files
	r.Static("/uploads", cfg.Upload.Dir)

	api := r.Group("/api/v1")
	{
		// Public auth endpoints (with rate limiting)
		auth := api.Group("/auth")
		{
			auth.POST("/login", middleware.RateLimit(10, 60*time.Second), authHandler.Login)
			auth.POST("/login/totp", middleware.RateLimit(10, 60*time.Second), authHandler.LoginWithTOTP)
			auth.POST("/passkey/login/options", middleware.RateLimit(10, 60*time.Second), authHandler.PasskeyLoginOptions)
			auth.POST("/passkey/login/verify", middleware.RateLimit(10, 60*time.Second), authHandler.PasskeyLoginVerify)
		}

		// Public config endpoint (for injection)
		api.GET("/config/public", siteConfigHandler.GetPublic)

		// Public content endpoints
		api.GET("/posts", postHandler.ListPublic)
		api.GET("/posts/top", postHandler.TopPosts)
		api.GET("/posts/top-viewed", postHandler.TopViewed)
		api.POST("/posts/:slug/reactions", postHandler.ToggleReaction)
		api.GET("/posts/:slug/adjacent", postHandler.AdjacentPosts)
		api.GET("/posts/:slug/related", postHandler.RelatedPosts)
		api.POST("/posts/:slug/verify-password", middleware.RateLimit(5, 60*time.Second), postHandler.VerifyPassword)
		api.GET("/posts/preview", postHandler.GetByPreviewToken)

		// Password reset (public, with rate limiting)
		api.POST("/auth/forgot-password", middleware.RateLimit(3, 60*time.Second), authHandler.ForgotPassword)
		api.POST("/auth/reset-password", middleware.RateLimit(5, 60*time.Second), authHandler.ResetPassword)
		api.GET("/posts/:slug", middleware.OptionalAuth(&cfg.JWT), postHandler.GetBySlug)
		api.GET("/archive", postHandler.Archive)

		// Friend links (public)
		api.GET("/links", friendLinkHandler.ListPublic)
		api.POST("/links/apply", middleware.RateLimit(5, 60*time.Second), friendLinkHandler.Apply)

			// Nav links (public)
			api.GET("/nav-links", navLinkHandler.ListPublic)

		api.GET("/categories", categoryHandler.List)
		api.GET("/categories/:slug", categoryHandler.GetBySlug)

		api.GET("/tags", tagHandler.List)
		api.GET("/tags/:slug", tagHandler.GetBySlug)

		api.GET("/series", seriesHandler.List)
		api.GET("/series/:slug", seriesHandler.GetBySlug)

		api.GET("/posts/:slug/comments", commentHandler.ListByPost)
		api.POST("/posts/:slug/comments", middleware.RateLimit(5, 60*time.Second), commentHandler.Create)
		api.POST("/posts/:slug/comments/:id/reactions", commentHandler.ToggleReaction)

		// Authenticated endpoints (CSRF protected)
		authRequired := api.Group("")
		authRequired.Use(middleware.AuthRequired(&cfg.JWT, db))
		authRequired.Use(middleware.CSRF())
		{
			// Auth management
			authRequired.POST("/auth/logout", authHandler.Logout)
			authRequired.GET("/auth/me", authHandler.Me)
			authRequired.PUT("/auth/profile", authHandler.UpdateProfile)
			authRequired.PUT("/auth/password", authHandler.ChangePassword)
			authRequired.POST("/auth/totp/setup", authHandler.TOTPSetup)
			authRequired.POST("/auth/totp/verify", authHandler.TOTPVerify)
			authRequired.DELETE("/auth/totp", authHandler.TOTPDisable)
			authRequired.POST("/auth/passkey/register/options", authHandler.PasskeyRegisterOptions)
			authRequired.POST("/auth/passkey/register/verify", authHandler.PasskeyRegisterVerify)
			authRequired.GET("/auth/passkeys", authHandler.ListPasskeys)
			authRequired.DELETE("/auth/passkey/:id", authHandler.DeletePasskey)
			authRequired.PUT("/auth/passkey/:id/rename", authHandler.RenamePasskey)

			// Admin: posts
			admin := authRequired.Group("/admin")
			admin.Use(middleware.RoleRequired("admin"))
			{
				admin.GET("/posts", postHandler.AdminList)
				admin.GET("/posts/:id", postHandler.AdminGet)
				admin.POST("/posts", postHandler.Create)
				admin.PUT("/posts/:id", postHandler.Update)
				admin.DELETE("/posts/:id", postHandler.Delete)
				admin.PATCH("/posts/:id/status", postHandler.UpdateStatus)
				admin.PATCH("/posts/:id/top", postHandler.ToggleTop)
				admin.GET("/posts/:id/revisions", postHandler.ListRevisions)
				admin.POST("/posts/:id/revisions/:revId/restore", postHandler.RestoreRevision)
				admin.POST("/posts/:id/preview-token", postHandler.GeneratePreviewToken)
				admin.GET("/posts/export", postHandler.Export)

				admin.POST("/categories", categoryHandler.Create)
				admin.PUT("/categories/:id", categoryHandler.Update)
				admin.DELETE("/categories/:id", categoryHandler.Delete)

				admin.POST("/tags", tagHandler.Create)
				admin.PUT("/tags/:id", tagHandler.Update)
				admin.DELETE("/tags/:id", tagHandler.Delete)

				admin.GET("/comments", commentHandler.AdminList)
				admin.PATCH("/comments/:id/status", commentHandler.UpdateStatus)
				admin.PATCH("/comments/batch-status", commentHandler.BatchUpdateStatus)
				admin.DELETE("/comments/:id", commentHandler.Delete)

				admin.POST("/upload", mediaHandler.Upload)
				admin.GET("/media", mediaHandler.List)
				admin.DELETE("/media/:id", mediaHandler.Delete)
				admin.PUT("/media/:id/tags", mediaHandler.UpdateMediaTags)
				admin.GET("/media/tags", mediaHandler.ListTags)
				admin.POST("/media/tags", mediaHandler.CreateTag)
				admin.DELETE("/media/tags/:id", mediaHandler.DeleteTag)

				admin.GET("/series", seriesHandler.AdminList)
				admin.POST("/series", seriesHandler.AdminCreate)
				admin.PUT("/series/:id", seriesHandler.AdminUpdate)
				admin.DELETE("/series/:id", seriesHandler.AdminDelete)
				admin.GET("/series/:id/posts", seriesHandler.AdminListPosts)
				admin.PUT("/series/:id/posts", seriesHandler.AdminSetPosts)

				admin.GET("/config", siteConfigHandler.Get)
				admin.PUT("/config", siteConfigHandler.Update)
				admin.POST("/config/test-email", siteConfigHandler.TestEmail)

				admin.GET("/users", authHandler.ListUsers)

				admin.GET("/access-logs", accessLogHandler.List)
				admin.GET("/access-logs/stats", accessLogHandler.Stats)
				admin.GET("/access-logs/export", accessLogHandler.Export)
				admin.DELETE("/access-logs/:id", accessLogHandler.Delete)
				admin.POST("/access-logs/clear", accessLogHandler.Clear)

				// Backup
				admin.GET("/backups", backupHandler.ListBackups)
				admin.POST("/backups", backupHandler.CreateBackup)
				admin.GET("/backups/:filename/download", backupHandler.DownloadBackup)
				admin.DELETE("/backups/:filename", backupHandler.DeleteBackup)

				// Restore
				admin.POST("/restore/upload", backupHandler.RestoreUpload)
				admin.POST("/restore/url", backupHandler.RestoreURL)
				admin.POST("/restore/local", backupHandler.RestoreLocal)

				// Friend links
				admin.GET("/links", friendLinkHandler.AdminList)
				admin.POST("/links", friendLinkHandler.AdminCreate)
				admin.PUT("/links/:id", friendLinkHandler.AdminUpdate)
				admin.PATCH("/links/:id/status", friendLinkHandler.AdminUpdateStatus)
				admin.DELETE("/links/:id", friendLinkHandler.AdminDelete)

					// Nav links
					admin.GET("/nav-links", navLinkHandler.AdminList)
					admin.POST("/nav-links", navLinkHandler.AdminCreate)
					admin.PUT("/nav-links/:id", navLinkHandler.AdminUpdate)
					admin.DELETE("/nav-links/:id", navLinkHandler.AdminDelete)
					admin.PUT("/nav-links/reorder", navLinkHandler.AdminReorder)
			}
		}
	}

	// RSS & Sitemap & Robots
	r.GET("/rss.xml", feedHandler.RSS)
	r.GET("/sitemap.xml", feedHandler.Sitemap)
	r.GET("/robots.txt", func(c *gin.Context) {
		siteURL := os.Getenv("SITE_URL")
		if siteURL == "" {
			siteURL = "https://tano.asia"
		}
		c.Header("Content-Type", "text/plain")
		c.String(http.StatusOK, "User-agent: *\nAllow: /\n\nSitemap: %s/sitemap.xml\n", siteURL)
	})

	// Health check with DB ping
	r.GET("/health", func(c *gin.Context) {
		sqlDB, err := db.DB()
		if err != nil || sqlDB.Ping() != nil {
			c.JSON(http.StatusServiceUnavailable, gin.H{"status": "error", "error": "database unreachable"})
			return
		}
		c.JSON(http.StatusOK, gin.H{"status": "ok"})
	})

	port := cfg.Server.Port
	srv := &http.Server{
		Addr:         ":" + port,
		Handler:      r,
		ReadTimeout:  cfg.Server.ReadTimeout,
		WriteTimeout: cfg.Server.WriteTimeout,
	}

	// Publish scheduled posts every minute
	go func() {
		ticker := time.NewTicker(1 * time.Minute)
		defer ticker.Stop()
		for range ticker.C {
			now := time.Now()
			result := db.Model(&model.Post{}).
				Where("status = ? AND published_at IS NOT NULL AND published_at <= ?", "draft", now).
				Updates(map[string]interface{}{"status": "published"})
			if result.RowsAffected > 0 {
				log.Printf("Published %d scheduled posts", result.RowsAffected)
			}
		}
	}()

	// Cleanup old backups every hour
	go func() {
		ticker := time.NewTicker(1 * time.Hour)
		defer ticker.Stop()
		for range ticker.C {
			backupHandler.CleanupOldBackups()
		}
	}()

	go func() {
		log.Printf("Server starting on port %s", port)
		if err := srv.ListenAndServe(); err != nil && err != http.ErrServerClosed {
			log.Fatalf("Failed to start server: %v", err)
		}
	}()

	quit := make(chan os.Signal, 1)
	signal.Notify(quit, syscall.SIGINT, syscall.SIGTERM)
	<-quit
	log.Println("Shutting down server...")

	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
	defer cancel()
	if err := srv.Shutdown(ctx); err != nil {
		log.Fatalf("Server forced to shutdown: %v", err)
	}
	log.Println("Server exited")
}

func seedAdmin(db *gorm.DB, password string) {
	var count int64
	db.Model(&model.User{}).Where("role = ?", "admin").Count(&count)
	if count > 0 {
		return
	}

	hash, err := utils.HashPassword(password)
	if err != nil {
		log.Fatalf("Failed to hash admin password: %v", err)
	}

	admin := model.User{
		Username:     "admin",
		Email:        "admin@tano.asia",
		PasswordHash: hash,
		DisplayName:  "管理员",
		Role:         "admin",
		Bio:          "A BanG Dreamer!",
	}

	if err := db.Create(&admin).Error; err != nil {
		log.Fatalf("Failed to seed admin user: %v", err)
	}

	log.Println("Default admin user created: admin")
}

func seedSiteConfigs(db *gorm.DB) {
	defaults := map[string]string{
		"site_title":       "朝花夕拾录",
		"site_description": "A BanG Dreamer!",
		"site_url":         "https://tano.asia",
		"footer_text":      "© 2026 Tano",
		"comment_enabled":  "true",
		"default_theme":    "dark",
		"accent_color":     "225",
		"email_enabled":    "false",
		"email_provider":   "zeabur",
		"email_from":       "",
		"profile_avatar":   "/aimi.png",
		"profile_name":     "Tano",
		"profile_bio":      "A BanG Dreamer!",
		"profile_contacts": `[{"type":"email","value":"public@tano.asia"},{"type":"github","value":"AhsokaTano26"}]`,
		"site_favicon":     "/favicon.ico",
	}

	for key, value := range defaults {
		var count int64
		db.Model(&model.SiteConfig{}).Where("key = ?", key).Count(&count)
		if count == 0 {
			db.Create(&model.SiteConfig{Key: key, Value: value})
		}
	}
}

func init() {
	_ = godotenv.Load()

	mode := os.Getenv("GIN_MODE")
	if mode == "" {
		mode = gin.ReleaseMode
	}
	gin.SetMode(mode)

	fmt.Println(`
  ╔══════════════════════════════════════╗
  ║        Tano Blog Backend API         ║
  ║        Powered by Go + Gin           ║
  ╚══════════════════════════════════════╝
	`)
}