# openapi.yaml
openapi: "3.0.3"
info:
  title: Tano Blog API
  description: 个人博客后端 API 接口
  version: "1.0.0"
  contact:
    name: Tano
    url: https://tano.asia
servers:
  - url: /api/v1
    description: API v1 基础路径
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 在 Authorization 头部使用 Bearer 令牌
    CookieAuth:
      type: apiKey
      in: cookie
      name: jwt
      description: 通过 jwt Cookie 传递令牌
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误描述
      required:
        - error
    Pagination:
      type: object
      properties:
        items:
          type: array
        total:
          type: integer
          description: 总数
        page:
          type: integer
          description: 当前页码
        size:
          type: integer
          description: 每页数量
...
    TokenResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserInfo'
    PostReactions:
      type: object
      additionalProperties:
        type: integer
      description: "表情符号到数量的映射，如 { \"👍\": 5 }"
    TOTPRequiredResponse:
      type: object
      properties:
        totp_required:
          type: boolean
          enum: [true]
        user_id:
          type: string
          format: uuid
...
    AdjacentPosts:
      type: object
      properties:
        prev:
          type: object
          nullable: true
          properties:
            slug:
              type: string
            title:
              type: string
        next:
          type: object
          nullable: true
          properties:
            slug:
              type: string
            title:
              type: string
...
paths:
...
  /auth/login:
    post:
      tags: [认证]
      summary: 用户登录
      description: 使用用户名和密码登录。如果用户启用了 TOTP，则需要继续调用 TOTP 验证接口。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  description: 用户名
                password:
                  type: string
                  description: 密码
                remember_me:
                  type: boolean
                  default: false
                  description: 记住登录状态（延长令牌有效期）
      responses:
        "200":
          description: 登录成功或需要 TOTP 二次验证
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TokenResponse'
                  - $ref: '#/components/schemas/TOTPRequiredResponse'
        "401":
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login/totp:
    post:
      tags: [认证]
      summary: TOTP 二次验证登录
      description: 当 Login 接口返回 totp_required=true 时，调用此接口完成 TOTP 验证登录。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, code]
              properties:
                user_id:
                  type: string
                  format: uuid
                  description: Login 接口返回的 user_id
                code:
                  type: string
                  description: TOTP 验证码（6位数字）
                remember_me:
                  type: boolean
                  default: false
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        "401":
          description: 验证码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/passkey/login/verify:
    post:
      tags: [认证]
      summary: 验证 Passkey 登录
      description: 验证前端 WebAuthn 凭证并返回 JWT 令牌。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: WebAuthn credential 对象的 JSON 序列化（rawBody）
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        "401":
          description: Passkey 验证失败
  /auth/forgot-password:
    post:
      tags: [密码重置]
      summary: 发送密码重置邮件
      description: 如果邮箱已注册，发送重置密码链接到该邮箱。无论邮箱是否存在均返回相同消息，防止枚举。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: 请求已处理
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "如果该邮箱已注册，重置链接将发送到您的邮箱"
  /auth/reset-password:
    post:
      tags: [密码重置]
      summary: 使用重置令牌重设密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                  description: 重置令牌（通过邮件接收）
                new_password:
                  type: string
                  minLength: 8
                  description: 新密码（至少8位）
      responses:
        "200":
          description: 密码已重置
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 密码已重置
        "400":
          description: 令牌无效或已过期

  # ========================================
  # Public Config
  # ========================================
...
  /posts:
    get:
      tags: [文章]
      summary: 获取已发布的文章列表
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
        - name: category
          in: query
          schema:
            type: string
          description: 按分类 slug 筛选
        - name: tag
          in: query
          schema:
            type: string
          description: 按标签 slug 筛选
        - name: search
          in: query
          schema:
            type: string
          description: 搜索关键词
      responses:
        "200":
          description: 文章列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pagination'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/PostItem'
...
  /posts/{slug}:
    get:
      tags: [文章]
      summary: 获取文章详情
      description: 已登录用户可查看草稿，匿名用户仅能看到已发布文章。返回内容包含文章反应统计。
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 文章详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  post:
                    type: object
                  reactions:
                    $ref: '#/components/schemas/PostReactions'
                  user_emojis:
                    type: array
                    items:
                      type: string
        "404":
          description: 文章不存在
  /posts/preview:
    get:
      tags: [文章]
      summary: 通过预览令牌获取草稿文章
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 文章详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  post:
                    type: object
  /posts/{slug}/reactions:
    post:
      tags: [文章]
      summary: 切换文章表情反应
      description: 添加或移除对文章的表情反应（基于 IP 去重）。
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emoji]
              properties:
                emoji:
                  type: string
                  description: 表情符号（1-10 个字符）
      responses:
        "200":
          description: 操作结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  active:
                    type: boolean
                    description: true=已添加, false=已移除
                  emoji:
                    type: string
  /posts/{slug}/adjacent:
    get:
      tags: [文章]
      summary: 获取上一篇/下一篇文章
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 相邻文章
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdjacentPosts'
  /posts/{slug}/related:
    get:
      tags: [文章]
      summary: 获取相关文章推荐
      description: 基于相同标签的文章推荐，最多返回 6 篇。
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 推荐文章列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostItem'
...
  /categories/{slug}:
    get:
      tags: [分类]
      summary: 获取分类详情及其文章列表
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        "200":
          description: 分类详情及文章
          content:
            application/json:
              schema:
                type: object
                properties:
                  category:
                    $ref: '#/components/schemas/Category'
                  posts:
                    type: array
                  total:
                    type: integer
                  page:
                    type: integer
                  size:
                    type: integer

  # ========================================
  # Public Content - Tags
  # ========================================
...
  /tags/{slug}:
    get:
      tags: [标签]
      summary: 获取标签详情及其文章列表
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        "200":
          description: 标签详情及文章
          content:
            application/json:
              schema:
                type: object
                properties:
                  tag:
                    $ref: '#/components/schemas/Tag'
                  posts:
                    type: array
                  total:
                    type: integer
                  page:
                    type: integer
                  size:
                    type: integer

  # ========================================
  # Public Content - Series
  # ========================================
...
  /series/{slug}:
    get:
      tags: [系列]
      summary: 获取系列详情及其文章
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 系列详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  series:
                    allOf:
                      - $ref: '#/components/schemas/Series'
                      - type: object
                        properties:
                          posts:
                            type: array
                            items:
                              $ref: '#/components/schemas/PostItem'

  # ========================================
  # Comments (Public)
  # ========================================
  /posts/{slug}/comments:
    get:
      tags: [评论]
      summary: 获取文章评论列表
      description: 返回树形结构的评论，包含表情反应统计。
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 评论列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'

    post:
      tags: [评论]
      summary: 创建评论
      description: 支持创建新评论和回复已有评论。包含蜜罐字段防止机器人。
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nickname, content]
              properties:
                parent_id:
                  type: string
                  description: 回复时的父评论 ID（可选）
                nickname:
                  type: string
                  maxLength: 100
                  description: 昵称
                email:
                  type: string
                  format: email
                  maxLength: 255
                  description: 邮箱（可选）
                website:
                  type: string
                  maxLength: 500
                  description: 网站地址（可选）
                content:
                  type: string
                  maxLength: 5000
                  description: 评论内容
                hp_field:
                  type: string
                  description: 蜜罐字段（留空即可）
      responses:
        "201":
          description: 评论已创建
          content:
            application/json:
              schema:
                type: object
                properties:
                  comment:
                    type: object
        "403":
          description: 该文章已关闭评论
  /posts/{slug}/comments/{id}/reactions:
    post:
      tags: [评论]
      summary: 切换评论表情反应
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emoji]
              properties:
                emoji:
                  type: string
      responses:
        "200":
          description: 操作结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  active:
                    type: boolean
                  emoji:
                    type: string

  # ========================================
  # Authenticated - Auth Management
  # ========================================
  /auth/logout:
    post:
      tags: [认证管理]
      summary: 退出登录
      description: 递增 token_version 使所有现有令牌失效。
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        "200":
          description: 已退出
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 已退出登录
  /auth/me:
    get:
      tags: [认证管理]
      summary: 获取当前用户信息
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        "200":
          description: 用户信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  username:
                    type: string
                  email:
                    type: string
                  display_name:
                    type: string
                  avatar_url:
                    type: string
                  bio:
                    type: string
                  role:
                    type: string
                  totp_enabled:
                    type: boolean
  /auth/profile:
    put:
      tags: [认证管理]
      summary: 更新个人资料
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                  maxLength: 100
                email:
                  type: string
                  format: email
                  maxLength: 255
                avatar_url:
                  type: string
                  maxLength: 500
                bio:
                  type: string
      responses:
        "200":
          description: 更新后的用户信息
          content:
            application/json:
              schema:
                type: object
  /auth/password:
    put:
      tags: [认证管理]
      summary: 修改密码
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [old_password, new_password]
              properties:
                old_password:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: 密码已修改
  /auth/totp/setup:
    post:
      tags: [认证管理]
      summary: 开始 TOTP 设置
      description: 生成 TOTP 密钥和 QR 码 URL，但尚未启用。
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        "200":
          description: TOTP 设置信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                    description: BASE32 密钥
                  qr_code:
                    type: string
                    description: otpauth:// URL（可生成 QR 码）
  /auth/totp/verify:
    post:
      tags: [认证管理]
      summary: 验证并启用 TOTP
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  description: TOTP 验证码
      responses:
        "200":
          description: TOTP 已启用
  /auth/totp:
    delete:
      tags: [认证管理]
      summary: 禁用 TOTP
      description: 需要提供密码或 TOTP 验证码之一来验证身份。
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: 当前密码
                totp_code:
                  type: string
                  description: TOTP 验证码
      responses:
        "200":
          description: TOTP 已禁用
        "400":
          description: 未提供密码或验证码
        "401":
          description: 密码错误或验证码错误
  /auth/passkey/register/options:
    post:
      tags: [认证管理]
      summary: 获取 Passkey 注册挑战
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        "200":
          description: WebAuthn 创建选项
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebAuthnOptions'
  /auth/passkey/register/verify:
    post:
      tags: [认证管理]
      summary: 验证并完成 Passkey 注册
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: WebAuthn credential 的 JSON 序列化（rawBody）
      responses:
        "200":
          description: Passkey 注册成功
...
  /auth/passkey/{id}:
    delete:
      tags: [认证管理]
      summary: 删除指定 Passkey
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 已删除
...